inference.py
#!/usr/bin/env python3
"""
Image Background Remover using BiRefNet
"""

import argparse
import numpy as np
from PIL import Image
import torch

def remove_background(input_path, output_path, model_path="BiRefNet"):
    """
    Remove background from an image and save the result
    """
    # TODO: Load BiRefNet model and run inference
    # This is a placeholder for the actual implementation
    print(f"Processing image: {input_path}")
    print(f"Output will be saved to: {output_path}")
    
    # Open input image
    img = Image.open(input_path).convert("RGBA")
    print(f"Input image size: {img.size}")
    
    # Placeholder: In a real implementation, you'd run BiRefNet here
    # For now, just copy the image with transparent background placeholder
    img.save(output_path)
    print(f"Saved output to {output_path}")
    return True

def main():
    parser = argparse.ArgumentParser(description='Remove background from an image')
    parser.add_argument('--input', required=True, help='Path to input image')
    parser.add_argument('--output', required=True, help='Path to output image (PNG)')
    parser.add_argument('--model', default='BiRefNet', help='Model path or name')
    args = parser.parse_args()
    
    success = remove_background(args.input, args.output, args.model)
    if not success:
        exit(1)

if __name__ == "__main__":
    main()
