requirements.txt
torch
transformers
pillow
numpy
argparse
